medical-bot/requirements.txt
flask
python-dotenv
langchain
langchain-pinecone
langchain-openai
langchain-community
langchain-core
langchain-huggingface
sentence-transformers==2.2.2
huggingface_hub==0.20.3
gTTS
pinecone
PineconeVectorStore
protobuf 
googleapis-common-protos
pinecone-client
pydub
SpeechRecognition
pyttsx3
transformers
torch
pypdf
